Safari-611.1.10/LayoutTests/http/tests/appcache/offline-access-expected.txt

CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Application Cache manifest could not be fetched, because a redirection was attempted.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
Test that offline applications work when there is no network access (simulated).

Should say SUCCESS:

SUCCESS
